Add link type reporting for pages

+++ SiteMapUriExtraction/SiteReporter.cs
@@ -34,7 +34,7 @@ namespace SiteMapUriExtractor {
             var backupPath = Path.Combine(outputFolder.FullName, backup);
             var filePath = Path.Combine(outputFolder.FullName, fileName);
 
-            if (File.Exists(backupPath)) {
+            if (File.Exists(backupPath) && File.Exists(filePath)) {
                 File.Delete(backupPath);
             }
             if (File.Exists(filePath)) {
@@ -44,29 +44,73 @@ namespace SiteMapUriExtractor {
             using var workBook = new XLWorkbook();
 
             var pagesSheet = workBook.Worksheets.Add("Pages");
-            var linksSheet = workBook.Worksheets.Add("Links");
+            var pagesData = GetPageData();
+            ReportPages(pagesSheet, pagesData);
+            PageData.FormatPages(pagesSheet);
+
+            var pageLinksSheet = workBook.Worksheets.Add("Page Links");
+
+            ReportPageLinks(pageLinksSheet, pagesData);
+            PageData.FormatPageLinks(pageLinksSheet);
+
+            var linksSheet = workBook.Worksheets.Add("All Links");
             ReportLinks(linksSheet);
-            ReportPages(pagesSheet);
-            RowData.Format(linksSheet);
-            PageData.Format(pagesSheet);
+            LinkData.Format(linksSheet);
             workBook.SaveAs(filePath);
 
         }
 
-        private void ReportPages(IXLWorksheet sheet) {
-            List<PageData> pageDatas = new List<PageData>();
-            foreach(Page page in pages) {
-                pageDatas.Add(new PageData(page));
+        private List<PageData> GetPageData() {
+            List<PageData> pagesData = new List<PageData>();
+            Dictionary<Uri, PageData> byUri = new();
+            foreach (Page page in pages) {
+                var data = new PageData(root, page);
+                pagesData.Add(data);
+                byUri.Add(page.Uri, data);
             }
-            PageData.WriteHeader(sheet);
+
+            foreach (Page page in pages) {
+                if (!byUri.TryGetValue(page.Uri, out var data)) {
+                    continue;
+                }
+                foreach(var reference in page.OutgoingReferences) {
+                    if (reference.HasTargetPage) {
+                        var targetPage = reference.TargetPage;
+                        if (!byUri.TryGetValue(targetPage.Uri, out var targetData)) {
+                            continue;
+                        }
+                        if (page.Uri == targetPage.Uri) {
+                            continue;
+                        }
+                        data.AddTarget(reference.ReferenceType, targetData);
+                    }
+                }
+            }
+
+            return pagesData;
+        }
+
+
+        private void ReportPages(IXLWorksheet sheet, List<PageData> pagesData) {
+            PageData.WritePageHeader(sheet);
             int row = 2;
-            foreach(var pageData in pageDatas) {
-                pageData.WriteRecord(sheet, row++);
+            foreach (var pageData in pagesData) {
+                pageData.WritePageRecord(sheet, row++);
+            }
+        }
+
+        private void ReportPageLinks(IXLWorksheet sheet, List<PageData> pagesData) {
+            PageData.WritePageLinksHeader(sheet);
+            int row = 2;
+            foreach (var pageData in pagesData) {
+                foreach (var referencedPage in pageData.ReferencedPages) {
+                    pageData.WritePageLinksRecord(sheet, row++, referencedPage);
+                }
             }
         }
 
         private void ReportLinks(IXLWorksheet sheet) {
-            RowData.WriteHeader(sheet);
+            LinkData.WriteHeader(sheet);
             int row = 2;
             foreach (Page page in pages) {
                 if (page.References == 0) {
@@ -79,12 +123,12 @@ namespace SiteMapUriExtractor {
         }
 
         private void WriteNotReferecedPage(IXLWorksheet sheet, Page page, int row) {
-            var data = new RowData(root, page);
+            var data = new LinkData(root, page);
             data.WriteRecord(sheet, row);
         }
 
         private void WriteReference(IXLWorksheet sheet, Reference reference, int row) {
-            var data = new RowData(root, reference);
+            var data = new LinkData(root, reference);
             data.WriteRecord(sheet, row);
         }
     }
